pdf_question_extractor/requirements.txt
# Core dependencies
fastapi==0.115.0
uvicorn[standard]==0.34.0
httpx==0.28.0
python-multipart==0.0.25

# API clients
mistralai==1.2.3
google-genai==0.9.0

# Processing
langchain-text-splitters==0.3.4
pydantic==2.10.4
tenacity==9.0.0

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
psycopg2-binary==2.9.10
pgvector==0.3.6
alembic==1.15.0

# Utilities
python-dotenv==1.0.1
aiofiles==24.1.0

# Development
pytest==8.3.4
pytest-asyncio==0.25.0